+++ source/app/components/molecules/row-item-template/styles.ts
@@ -1,0 +1,9 @@
+export default {
+  textFolder: {
+    color: 'text.third',
+    paddingLeft: '16px'
+  },
+  buttonSelect: {
+    mt: '3px'
+  }
+};