+++ source/src/view/ViewLoader.ts
@@ -15,16 +15,23 @@ export class ViewLoader {
     this.context = context;
     this.disposables = [];
 
-    this.panel = vscode.window.createWebviewPanel('celerikScaffolder', 'Celerik Scaffolder', vscode.ViewColumn.One, {
-      enableScripts: true,
-      retainContextWhenHidden: true,
-      localResourceRoots: [vscode.Uri.file(path.join(this.context.extensionPath, 'out', 'app'))]
-    });
+    this.panel = vscode.window.createWebviewPanel(
+      'celerikScaffolder',
+      'Celerik Scaffolder',
+      vscode.ViewColumn.One,
+      {
+        enableScripts: true,
+        retainContextWhenHidden: true,
+        localResourceRoots: [vscode.Uri.file(path.join(this.context.extensionPath, 'out', 'app'))]
+      }
+    );
     const templateUrl = 'global.state';
     const state = context.globalState.get(templateUrl);
     if (!state) {
       // Add a initial url value
-      const data = JSON.stringify({ templateUrl: 'https://github.com/celerik/celerik-scaffolder-templates.git' });
+      const data = JSON.stringify({
+        templateUrl: 'https://github.com/celerik/celerik-scaffolder-templates.git'
+      });
       context.globalState.update(templateUrl, data);
     }
 
@@ -41,7 +48,10 @@ export class ViewLoader {
           context.globalState.update(templateUrl, JSON.stringify(text));
         } else if (message.type === 'COMMON') {
           const text = (message as CommonMessage).payload;
-          vscode.window.showInformationMessage(`Received message from Webview: ${text}`);
+          vscode.window.showInformationMessage(`${text}`);
+        } else if (message.type === 'ERROR') {
+          const text = (message as CommonMessage).payload;
+          vscode.window.showErrorMessage(`${text}`);
         }
       },
       null,
@@ -101,7 +111,7 @@ export class ViewLoader {
 
     // The global status of vs code is loaded and passed as a string to the webview.
     let prevState = this.context.globalState.get('global.state') || '';
-    prevState = JSON.stringify(prevState).replace(/\\"/g, '\'');
+    prevState = JSON.stringify(prevState).replace(/\\"/g, "'");
 
     return `
       <!DOCTYPE html>